Adds initial parser implementation
The implementation is very basic. Especially error handling and error
messages need more attention and functionality. The core functionality
of parsing lexed tokens into a syntax tree is mostly tested and seems
to be working as intended.

The use of Result in general is something else to pay attention
to. Without proper type inference for lambdas directly, the existing
API for Result is cumbersome - something better should be
introduced (probably as one of the first tasks of phase 2).