Adds executable generation
Very naive implementation of end-to-end executable generation and
testing. This allows Jackal's code generator to persist executable
files to disk which the user would be able to interact with directly.

With this implemented, the compiler is now technically "end-to-end" in
that it can take Jackal source code and generate executable machine
code.

After further testing, phase 1 of the compiler will be effectively
complete, though there are a few TODOs that I may want to address
before tagging phase 1.