Checkpoint: further implementation of V1 parser and ast
Parser is still non-functional, but the introduction of top-level
forms as the primary means by which overall programs will be
structured provides a simple way for parsing to be implemented and
tested more incrementally. The first increment will include only the
Function and Data forms. This will allow us to get a simplified
version of the V1 language functional before we switch focus to the
higher-order and more difficult to implement features of the basic language.